+++ config/server.js
@@ -1,7 +1,8 @@
 const express = require('express');
-const router = require('./config/authRoute');
 const cors = require('cors');
 const helmet = require('helmet');
+const usersRouter = require('./usersRouter');
+const challengesRouter = require('./challengesRouter');
 
 const server = express();
 
@@ -12,6 +13,7 @@ server.get('/', (req, res) =>
     res.status(200).send({
         message: "API is alive"
     }));
-server.use('/users', router);
+server.use('/users', usersRouter);
+server.use('/challenges', challengesRouter);
 
 module.exports = server;
