+++ tests/apis/db/usersDb.test.js
@@ -1,5 +1,5 @@
-const db = require('../data/dbConfig');
-const dbApi = require('../api/usersDb');
+const db = require('../../../data/dbConfig');
+const usersApi = require('../../../apis/db/users');
 
 describe('Users Model', () => {
     beforeEach(async () => {
@@ -8,7 +8,7 @@ describe('Users Model', () => {
 
     describe('insert()', () => {
         it('should insert user', async () => {
-            const userId = await dbApi.insert({
+            const userId = await usersApi.insert({
                 sub_id: 'auth0sub'
             });
             const user = await db('users').where({
@@ -20,10 +20,10 @@ describe('Users Model', () => {
 
     describe('getOne()', () => {
         it('should get a user by their sub_id', async () => {
-            const userId = await dbApi.insert({
+            const userId = await usersApi.insert({
                 sub_id: 'auth0sub'
             });
-            const user = await dbApi.getOne({sub_id: userId.sub_id});
+            const user = await usersApi.getOne({sub_id: userId.sub_id});
 
             expect(user.sub_id).toBe('auth0sub');
         });
@@ -31,14 +31,14 @@ describe('Users Model', () => {
 
     describe('getMany()', () => {
         it('should get all users', async () => {
-            await dbApi.insert([{
+            await usersApi.insert([{
                 sub_id: 'auth0sub1'
             }, {
                 sub_id: 'auth0sub2'
             }, {
                 sub_id: 'auth0sub3'
             }]);
-            const users = await dbApi.getMany()
+            const users = await usersApi.getMany()
             expect(users).toHaveLength(3);
         });
     });
